Parse CSB date

csb_date_parse is used to parse out dates into more intellgible formats

csb_date_parse(.data, var, day, month, year, delete = FALSE)

Arguments

.data
A tibble with raw CSB data

var
name of column containing date data

day
if specified, returns a named column with parsed day

month
if specified, returns a named column with parsed month

year
if specified, returns a named column with parsed year

delete
if true, deletes the original column with unparsed data

Value

csb_date_parse returns new columns with specified names containing parsed date information

Examples

csb_date_parse(january_2018, DATETIMEINIT, dayInit)

#> # A tibble: 1,688 x 20
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  2 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  3 WEB        NA                  NA                  NA                 
#>  4 WEB        NA                  2018-01-05 16:17:41 2018-01-05 16:18:08
#>  5 PHONE      NA                  2018-01-01 10:40:14 2018-01-01 10:40:54
#>  6 WEB        NA                  2018-01-02 12:54:41 2018-01-12 10:27:12
#>  7 PHONE      NA                  2018-01-03 11:22:29 2018-01-03 11:22:29
#>  8 PHONE      NA                  2018-01-03 11:19:13 2018-01-03 11:19:13
#>  9 WEB        NA                  2018-01-05 16:16:55 2018-01-05 16:17:28
#> 10 PHONE      NA                  2017-12-29 20:00:00 2018-01-01 15:01:19
#> # … with 1,678 more rows, and 16 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>,
#> #   SUBMITTO <chr>, WARD <int>, dayInit <int>

csb_date_parse(january_2018, DATETIMEINIT, dayInit, monthInit)

#> # A tibble: 1,688 x 21
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  2 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  3 WEB        NA                  NA                  NA                 
#>  4 WEB        NA                  2018-01-05 16:17:41 2018-01-05 16:18:08
#>  5 PHONE      NA                  2018-01-01 10:40:14 2018-01-01 10:40:54
#>  6 WEB        NA                  2018-01-02 12:54:41 2018-01-12 10:27:12
#>  7 PHONE      NA                  2018-01-03 11:22:29 2018-01-03 11:22:29
#>  8 PHONE      NA                  2018-01-03 11:19:13 2018-01-03 11:19:13
#>  9 WEB        NA                  2018-01-05 16:16:55 2018-01-05 16:17:28
#> 10 PHONE      NA                  2017-12-29 20:00:00 2018-01-01 15:01:19
#> # … with 1,678 more rows, and 17 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>,
#> #   SUBMITTO <chr>, WARD <int>, dayInit <int>, monthInit <dbl>

csb_date_parse(january_2018, DATETIMEINIT, month = monthInit)

#> # A tibble: 1,688 x 20
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  2 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  3 WEB        NA                  NA                  NA                 
#>  4 WEB        NA                  2018-01-05 16:17:41 2018-01-05 16:18:08
#>  5 PHONE      NA                  2018-01-01 10:40:14 2018-01-01 10:40:54
#>  6 WEB        NA                  2018-01-02 12:54:41 2018-01-12 10:27:12
#>  7 PHONE      NA                  2018-01-03 11:22:29 2018-01-03 11:22:29
#>  8 PHONE      NA                  2018-01-03 11:19:13 2018-01-03 11:19:13
#>  9 WEB        NA                  2018-01-05 16:16:55 2018-01-05 16:17:28
#> 10 PHONE      NA                  2017-12-29 20:00:00 2018-01-01 15:01:19
#> # … with 1,678 more rows, and 16 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>,
#> #   SUBMITTO <chr>, WARD <int>, monthInit <dbl>

csb_date_parse(january_2018, DATETIMEINIT, month = monthInit, year = yearInit)

#> # A tibble: 1,688 x 21
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  2 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  3 WEB        NA                  NA                  NA                 
#>  4 WEB        NA                  2018-01-05 16:17:41 2018-01-05 16:18:08
#>  5 PHONE      NA                  2018-01-01 10:40:14 2018-01-01 10:40:54
#>  6 WEB        NA                  2018-01-02 12:54:41 2018-01-12 10:27:12
#>  7 PHONE      NA                  2018-01-03 11:22:29 2018-01-03 11:22:29
#>  8 PHONE      NA                  2018-01-03 11:19:13 2018-01-03 11:19:13
#>  9 WEB        NA                  2018-01-05 16:16:55 2018-01-05 16:17:28
#> 10 PHONE      NA                  2017-12-29 20:00:00 2018-01-01 15:01:19
#> # … with 1,678 more rows, and 17 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>,
#> #   SUBMITTO <chr>, WARD <int>, monthInit <dbl>, yearInit <dbl>

csb_date_parse(january_2018, DATETIMEINIT, dayInit, monthInit, yearInit, delete = TRUE)

#> # A tibble: 1,688 x 21
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  2 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  3 WEB        NA                  NA                  NA                 
#>  4 WEB        NA                  2018-01-05 16:17:41 2018-01-05 16:18:08
#>  5 PHONE      NA                  2018-01-01 10:40:14 2018-01-01 10:40:54
#>  6 WEB        NA                  2018-01-02 12:54:41 2018-01-12 10:27:12
#>  7 PHONE      NA                  2018-01-03 11:22:29 2018-01-03 11:22:29
#>  8 PHONE      NA                  2018-01-03 11:19:13 2018-01-03 11:19:13
#>  9 WEB        NA                  2018-01-05 16:16:55 2018-01-05 16:17:28
#> 10 PHONE      NA                  2017-12-29 20:00:00 2018-01-01 15:01:19
#> # … with 1,678 more rows, and 17 more variables: DESCRIPTION <chr>,
#> #   NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>, PROBADDRESS <chr>,
#> #   PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>, PROBZIP <int>,
#> #   REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>, SUBMITTO <chr>,
#> #   WARD <int>, dayInit <int>, monthInit <dbl>, yearInit <dbl>