CSB intersections

csb_intersectionreturns a logical vector indicating `TRUE` for addresses that are intersections. This function is also able to isolate or remove these data.

csb_intersection(.data, var, newVar, filter = FALSE, remove = FALSE)

Arguments

.data
A tibble with raw CSB data

var
name of column containing address

newVar
Name of output variable to be created with logical

filter
If true, returns a filtered version of the input with only observations occurring at intersections

remove
If true, returns a version of the input without observations occuring at intersections

Value

csb_intersectionreturns a logical vector for addresses that are intersections, or it filters intersections

Examples

# NOT RUN {
csb_intersection(january_2018, PROBADDRESS, newVar = Intersection)
# }
csb_intersection(january_2018, PROBADDRESS, filter = TRUE)

#> No argument specified for newVar, no logical will be appended

#> 1454 observations were filtered out

#> # A tibble: 234 x 19
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 PHONE      NA                  2018-01-01 10:40:14 2018-01-01 10:40:54
#>  2 PHONE      NA                  2018-01-03 11:19:13 2018-01-03 11:19:13
#>  3 WEB        NA                  2018-01-05 16:16:55 2018-01-05 16:17:28
#>  4 PHONE      NA                  2017-12-30 01:15:00 2018-01-01 14:52:00
#>  5 WEB        NA                  2018-03-06 09:11:48 2018-03-06 09:11:48
#>  6 PHONE      NA                  2018-01-02 11:28:34 2018-01-02 11:28:34
#>  7 PHONE      NA                  2018-01-02 11:50:00 2018-01-02 16:37:50
#>  8 WEB        NA                  2018-01-05 16:10:39 2018-01-05 16:11:03
#>  9 WEB        NA                  2018-01-03 16:56:06 2018-01-03 16:48:12
#> 10 PHONE      NA                  2018-01-02 08:39:07 2018-01-02 08:39:07
#> # … with 224 more rows, and 15 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>,
#> #   SUBMITTO <chr>, WARD <int>

csb_intersection(january_2018, PROBADDRESS, remove = TRUE)

#> No argument specified for newVar, no logical will be appended

#> 248 observations were removed

#> # A tibble: 1,440 x 19
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  2 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  3 WEB        NA                  NA                  NA                 
#>  4 WEB        NA                  2018-01-05 16:17:41 2018-01-05 16:18:08
#>  5 WEB        NA                  2018-01-02 12:54:41 2018-01-12 10:27:12
#>  6 PHONE      NA                  2018-01-03 11:22:29 2018-01-03 11:22:29
#>  7 PHONE      NA                  2017-12-29 20:00:00 2018-01-01 15:01:19
#>  8 PHONE      NA                  2018-01-04 15:39:56 2018-01-04 15:39:56
#>  9 NA         2018-01-02 11:33:56 NA                  NA                 
#> 10 NA         2018-01-02 13:54:06 NA                  NA                 
#> # … with 1,430 more rows, and 15 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>,
#> #   SUBMITTO <chr>, WARD <int>