Load CSB Variable Definitions

Provides direct access to the CSB's variable definitions, which are available for download from the City of St. Louis's open data site.

csb_load_variables(tidy = TRUE)

Arguments

tidy
A logical scalar; if TRUE, variable names will be converted to lower case and reordered. Two variables with incomplete data - problem city (PROBCITY) and problem zip code (PROBZIP) - are dropped to limit use of memory. This mirrors the functionality of csb_get_data.

Value

A tibble containing variable names and definitions.

See also

Examples

csb_load_variables()

#> # A tibble: 17 x 2
#>    Name          Description                                                    
#>    <chr>         <chr>                                                          
#>  1 requestid     System generated unique record number                          
#>  2 datetimeinit  Date/Time the request was initiated (auto stamped based on whe…
#>  3 probaddress   The address where the problem is occurring                     
#>  4 probaddtype   A = Parcel, B = Intersection                                   
#>  5 callertype    Method used by customer to report this issue (Phone, Web, Twit…
#>  6 neighborhood  City of St. Louis Neighborhoods Number (1-79)                  
#>  7 ward          City of St. Louis Wards Number (1-28)                          
#>  8 problemcode   Type of report                                                 
#>  9 description   Same as problem code OR more specific explanation of problem c…
#> 10 submitto      City division that is responsible for completing this request  
#> 11 status        Status of the request                                          
#> 12 dateinvtdone  Date Investigation Done - the date the inspection/work was don…
#> 13 datetimeclos… Date/Time the request was closed (auto stamped based on when e…
#> 14 prjcompleted… Projected Completion Date - the date by which the SubmitTo div…
#> 15 datecancelled Date Cancelled - indicates the request was a duplicate, cancel…
#> 16 srx           Service Request X-Coordinate (map coordinate)                  
#> 17 sry           Service Request Y-Coordinate (map coordinate)                  

csb_load_variables(tidy = FALSE)

#> # A tibble: 19 x 2
#>    Name          Description                                                    
#>    <chr>         <chr>                                                          
#>  1 CALLERTYPE    Method used by customer to report this issue (Phone, Web, Twit…
#>  2 DATECANCELLED Date Cancelled - indicates the request was a duplicate, cancel…
#>  3 DATEINVTDONE  Date Investigation Done - the date the inspection/work was don…
#>  4 DATETIMECLOS… Date/Time the request was closed (auto stamped based on when e…
#>  5 DATETIMEINIT  Date/Time the request was initiated (auto stamped based on whe…
#>  6 DESCRIPTION   Same as problem code OR more specific explanation of problem c…
#>  7 NEIGHBORHOOD  City of St. Louis Neighborhoods Number (1-79)                  
#>  8 PRJCOMPLETED… Projected Completion Date - the date by which the SubmitTo div…
#>  9 PROBADDRESS   The address where the problem is occurring                     
#> 10 PROBADDTYPE   A = Parcel, B = Intersection                                   
#> 11 PROBCITY      The city where the problem is occurring (often blank, always C…
#> 12 PROBLEMCODE   Type of report                                                 
#> 13 PROBZIP       The zip code where the problem is occurring (often blank, not …
#> 14 REQUESTID     System generated unique record number                          
#> 15 SRX           Service Request X-Coordinate (map coordinate)                  
#> 16 SRY           Service Request Y-Coordinate (map coordinate)                  
#> 17 STATUS        Status of the request                                          
#> 18 SUBMITTO      City division that is responsible for completing this request  
#> 19 WARD          City of St. Louis Wards Number (1-28)