Project Calls for Service Data Using Coordinates

csb_projectXY converts SRX and SRY data into a simple features object. You can write a shapefile directly from the output of this function using sf::st_write.

csb_projectXY(.data, varX, varY, crs)

Arguments

.data
A tbl or data frame

varX
Name of column containing x coordinate data

varY
Name of column containing y coordinate data

crs
Optional; coordinate reference system for the data to be projected into

Value

Returns a sf object of the input data projected as point data.

Examples

# remove missing coordinates prior to projecting
csb <- csb_missingXY(january_2018, SRX, SRY, newVar = missing)
csb <- dplyr::filter(csb, missing == FALSE)

# project data
csb_projectXY(csb, SRX, SRY)

#> Simple feature collection with 1506 features and 18 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 872868.1 ymin: 985130.7 xmax: 910286.8 ymax: 1061922
#> epsg (SRID):    NA
#> proj4string:    +proj=tmerc +lat_0=35.83333333333334 +lon_0=-90.5 +k=0.9999333333333333 +x_0=250000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=us-ft +no_defs
#> # A tibble: 1,506 x 19
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  2 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  3 WEB        NA                  NA                  NA                 
#>  4 WEB        NA                  2018-01-05 16:17:41 2018-01-05 16:18:08
#>  5 PHONE      NA                  2018-01-01 10:40:14 2018-01-01 10:40:54
#>  6 WEB        NA                  2018-01-02 12:54:41 2018-01-12 10:27:12
#>  7 PHONE      NA                  2018-01-03 11:22:29 2018-01-03 11:22:29
#>  8 PHONE      NA                  2018-01-03 11:19:13 2018-01-03 11:19:13
#>  9 WEB        NA                  2018-01-05 16:16:55 2018-01-05 16:17:28
#> 10 PHONE      NA                  2017-12-29 20:00:00 2018-01-01 15:01:19
#> # … with 1,496 more rows, and 15 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, STATUS <chr>, SUBMITTO <chr>, WARD <int>,
#> #   missing <lgl>, geometry <POINT [US_survey_foot]>

# project with a custom crs
csb_projectXY(csb, SRX, SRY, crs = 4269)

#> Simple feature collection with 1506 features and 18 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -90.31572 ymin: 38.53886 xmax: -90.1846 ymax: 38.74956
#> epsg (SRID):    4269
#> proj4string:    +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs
#> # A tibble: 1,506 x 19
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  2 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  3 WEB        NA                  NA                  NA                 
#>  4 WEB        NA                  2018-01-05 16:17:41 2018-01-05 16:18:08
#>  5 PHONE      NA                  2018-01-01 10:40:14 2018-01-01 10:40:54
#>  6 WEB        NA                  2018-01-02 12:54:41 2018-01-12 10:27:12
#>  7 PHONE      NA                  2018-01-03 11:22:29 2018-01-03 11:22:29
#>  8 PHONE      NA                  2018-01-03 11:19:13 2018-01-03 11:19:13
#>  9 WEB        NA                  2018-01-05 16:16:55 2018-01-05 16:17:28
#> 10 PHONE      NA                  2017-12-29 20:00:00 2018-01-01 15:01:19
#> # … with 1,496 more rows, and 15 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, STATUS <chr>, SUBMITTO <chr>, WARD <int>,
#> #   missing <lgl>, geometry <POINT [°]>