CSB vacant codes

csb_vacantappends a logical vector indicating `TRUE` for vacancy related problem codes.

csb_vacant(.data, var, newVar, filter = FALSE)

Arguments

.data
A tibble with raw CSB data

var
name of column containing problem codes

newVar
name of output variable to be created with vacant logical

filter
If true, returns a filtered version of the input with only vacancy related problem codes

Value

csb_vacantreturns a logical vector with TRUE for vacancy related problem codes, or a filtered version of the input data

Examples

csb_vacant(january_2018, PROBLEMCODE, vacant)

#> # A tibble: 1,688 x 20
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  2 WEB        NA                  2018-01-05 16:22:19 2018-01-05 16:22:21
#>  3 WEB        NA                  NA                  NA                 
#>  4 WEB        NA                  2018-01-05 16:17:41 2018-01-05 16:18:08
#>  5 PHONE      NA                  2018-01-01 10:40:14 2018-01-01 10:40:54
#>  6 WEB        NA                  2018-01-02 12:54:41 2018-01-12 10:27:12
#>  7 PHONE      NA                  2018-01-03 11:22:29 2018-01-03 11:22:29
#>  8 PHONE      NA                  2018-01-03 11:19:13 2018-01-03 11:19:13
#>  9 WEB        NA                  2018-01-05 16:16:55 2018-01-05 16:17:28
#> 10 PHONE      NA                  2017-12-29 20:00:00 2018-01-01 15:01:19
#> # … with 1,678 more rows, and 16 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>,
#> #   SUBMITTO <chr>, WARD <int>, vacant <lgl>

csb_vacant(january_2018, PROBLEMCODE, vacant, filter = TRUE)

#> # A tibble: 98 x 20
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 PHONE      NA                  2018-01-02 13:59:00 2018-04-03 08:36:15
#>  2 PHONE      NA                  2018-01-29 13:19:59 2018-02-02 12:13:31
#>  3 WEB        NA                  2018-01-23 11:25:17 2018-01-23 11:25:20
#>  4 PHONE      NA                  2018-01-03 00:00:00 2018-01-11 15:13:08
#>  5 WEB        NA                  2018-01-03 08:15:36 2018-03-08 11:33:56
#>  6 PHONE      NA                  2018-01-22 15:37:57 2018-01-22 15:38:20
#>  7 PHONE      NA                  2018-01-12 14:05:32 2018-01-12 14:06:14
#>  8 PHONE      NA                  2018-01-04 15:58:00 2018-05-07 10:06:08
#>  9 PHONE      NA                  2018-01-02 14:09:00 2018-05-08 08:16:30
#> 10 WEB        NA                  2018-01-03 08:18:52 2018-05-07 10:53:27
#> # … with 88 more rows, and 16 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>,
#> #   SUBMITTO <chr>, WARD <int>, vacant <lgl>