CSB Calls in St. Louis, January 2018

An example data set containing CSB calls in St. Louis, Missouri during the first week of January, 2018.

data(january_2018)

Format

A tibble with 1554 rows and 17 variables:

requestid

system generated unique request identifier

datetimeinit

date and time the request was initiated

probaddress

address of the request

probaddtype

A = Parcel, B = Intersection

callertype

method used by citizen to report issue (Phone, Web, Twitter, etc)

neighborhood

City of St. Louis Neighborhood number (1-79)

ward

City of St. Louis Ward number (1-28)

problemcode

type of report

description

same as problemcode OR slightly more specific

submitto

city division responsible for completing the request

status

status of the request

dateinvtdone

date of investigation-date that work was done, may differ from closing date because of crews using paper copies of requests

datetimeclosed

date and time the request was closed

projcompletedate

date by which city division should have initial inspection complete, auto-populated based on service level agreements

datecancelled

indicates a duplicate, cancelled or entered in error request

srx

map coordinate, X-coordinate

sry

map coordinate, Y-coordinate

Source

St. Louis Citizens' Service Bureau

Examples

str(january_2018)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1554 obs. of  17 variables:
#>  $ requestid      : num  1049265 1049265 1049266 1049267 1049268 ...
#>  $ datetimeinit   : POSIXct, format: "2018-01-01 00:48:30" "2018-01-01 00:48:30" ...
#>  $ probaddress    : chr  "75 WILLMORE RD" "75 WILLMORE RD" "6739 VERMONT AVE" "5444 ITASKA ST" ...
#>  $ probaddtype    : chr  "A" "A" "A" "A" ...
#>  $ callertype     : chr  "WEB" "WEB" "WEB" "WEB" ...
#>  $ neighborhood   : num  8 8 1 7 60 42 19 25 25 79 ...
#>  $ ward           : num  16 16 11 16 5 24 20 8 6 2 ...
#>  $ problemcode    : chr  "WTR-COMING-UP" "WTR-COMING-UP" "NOWATER" "NOWATER" ...
#>  $ description    : chr  "WTR-COMING-UP" "WTR-COMING-UP" "NOWATER (due to main break, etc.)" "NOWATER (due to main break, etc.)" ...
#>  $ submitto       : chr  "WATER DIV," "WATER DIV," "WATER DIV," "WATER DIV," ...
#>  $ status         : chr  "CLOSED" "CLOSED" "Disptch" "CLOSED" ...
#>  $ dateinvtdone   : POSIXct, format: "2018-01-05 16:22:19" "2018-01-05 16:22:19" ...
#>  $ datetimeclosed : POSIXct, format: "2018-01-05 16:22:21" "2018-01-05 16:22:21" ...
#>  $ prjcompletedate: POSIXct, format: "2018-01-15" "2018-01-15" ...
#>  $ datecancelled  : POSIXct, format: NA NA ...
#>  $ srx            : num  875855 875855 889805 881255 904418 ...
#>  $ sry            : num  998435 998435 991410 1001625 1023278 ...
#>  - attr(*, "spec")=
#>   .. cols(
#>   ..   requestid = col_double(),
#>   ..   datetimeinit = col_datetime(format = ""),
#>   ..   probaddress = col_character(),
#>   ..   probaddtype = col_character(),
#>   ..   callertype = col_character(),
#>   ..   neighborhood = col_double(),
#>   ..   ward = col_double(),
#>   ..   problemcode = col_character(),
#>   ..   description = col_character(),
#>   ..   submitto = col_character(),
#>   ..   status = col_character(),
#>   ..   dateinvtdone = col_datetime(format = ""),
#>   ..   datetimeclosed = col_datetime(format = ""),
#>   ..   prjcompletedate = col_datetime(format = ""),
#>   ..   datecancelled = col_datetime(format = ""),
#>   ..   srx = col_double(),
#>   ..   sry = col_double()
#>   .. )

head(january_2018)

#> # A tibble: 6 x 17
#>   requestid datetimeinit        probaddress probaddtype callertype neighborhood
#>       <dbl> <dttm>              <chr>       <chr>       <chr>             <dbl>
#> 1   1049265 2018-01-01 00:48:30 75 WILLMOR… A           WEB                   8
#> 2   1049265 2018-01-01 00:48:30 75 WILLMOR… A           WEB                   8
#> 3   1049266 2018-01-01 05:29:06 6739 VERMO… A           WEB                   1
#> 4   1049267 2018-01-01 08:53:07 5444 ITASK… A           WEB                   7
#> 5   1049268 2018-01-01 09:13:22 CASS AVE &… B           PHONE                60
#> 6   1049269 2018-01-01 10:12:26 6441 LLOYD… A           WEB                  42
#> # … with 11 more variables: ward <dbl>, problemcode <chr>, description <chr>,
#> #   submitto <chr>, status <chr>, dateinvtdone <dttm>, datetimeclosed <dttm>,
#> #   prjcompletedate <dttm>, datecancelled <dttm>, srx <dbl>, sry <dbl>