Filter CSB Data

csb_filter retuns observations that match any combination of 17 predefined categories.

csb_filter(.data, var, newVar, category)

Arguments

.data
A tibble with raw CSB data

var
name of the column containg original problem code data

newVar
name of column containing categories. NULL by default, but if specified will produce a new column.

category
a vector with the unquoted name(s) of the category(s) for the function to return. You can also explicitly state quoted PROBLEMCODE(s). Valid categories are: (admin, animal, construction, debris, degrade, disturbance, event, health, landscape, law, maintenance, nature, road, sewer, traffic, vacant, waste)

Value

csb_filter returns data with an additional variable for an intelligible category for CSB requests.

Examples

csb_filter(january_2018, PROBLEMCODE, Category, category = c(cat_waste, cat_debris))

#> # A tibble: 521 x 20
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 WEB        NA                  2018-01-02 12:54:41 2018-01-12 10:27:12
#>  2 WEB        NA                  2018-01-04 12:55:01 2018-01-08 09:36:27
#>  3 WEB        NA                  2018-01-03 12:53:39 2018-01-05 12:24:22
#>  4 PHONE      NA                  2018-01-04 08:37:22 2018-01-04 08:37:22
#>  5 WEB        NA                  2018-01-03 12:54:29 2018-01-23 13:27:12
#>  6 PHONE      NA                  2018-01-02 13:59:00 2018-04-03 08:36:15
#>  7 PHONE      NA                  2018-01-02 09:55:07 2018-01-02 10:01:18
#>  8 TWITTER    NA                  2018-01-02 12:32:10 2018-01-02 12:32:10
#>  9 TWITTER    NA                  2018-01-09 07:41:00 2018-03-08 07:42:20
#> 10 PHONE      NA                  2018-01-06 10:55:07 2018-01-06 10:55:07
#> # … with 511 more rows, and 16 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>,
#> #   SUBMITTO <chr>, WARD <int>, Category <chr>

csb_filter(january_2018, PROBLEMCODE, category = cat_vacant)

#> No argument set for `newVar` If you would like to append a category variable, please set an argument for `newVar`

#> # A tibble: 98 x 19
#>    CALLERTYPE DATECANCELLED       DATEINVTDONE        DATETIMECLOSED     
#>    <chr>      <dttm>              <dttm>              <dttm>             
#>  1 PHONE      NA                  2018-01-02 13:59:00 2018-04-03 08:36:15
#>  2 PHONE      NA                  2018-01-29 13:19:59 2018-02-02 12:13:31
#>  3 WEB        NA                  2018-01-23 11:25:17 2018-01-23 11:25:20
#>  4 PHONE      NA                  2018-01-03 00:00:00 2018-01-11 15:13:08
#>  5 WEB        NA                  2018-01-03 08:15:36 2018-03-08 11:33:56
#>  6 PHONE      NA                  2018-01-22 15:37:57 2018-01-22 15:38:20
#>  7 PHONE      NA                  2018-01-12 14:05:32 2018-01-12 14:06:14
#>  8 PHONE      NA                  2018-01-04 15:58:00 2018-05-07 10:06:08
#>  9 PHONE      NA                  2018-01-02 14:09:00 2018-05-08 08:16:30
#> 10 WEB        NA                  2018-01-03 08:18:52 2018-05-07 10:53:27
#> # … with 88 more rows, and 15 more variables: DATETIMEINIT <dttm>,
#> #   DESCRIPTION <chr>, NEIGHBORHOOD <int>, PRJCOMPLETEDATE <dttm>,
#> #   PROBADDRESS <chr>, PROBADDTYPE <chr>, PROBCITY <chr>, PROBLEMCODE <chr>,
#> #   PROBZIP <int>, REQUESTID <int>, SRX <dbl>, SRY <dbl>, STATUS <chr>,
#> #   SUBMITTO <chr>, WARD <int>